+++ Assets/Scripts/CustomEditors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 697e53ba7aac2f642b213c4d16979bd0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
