+++ Assets/Scripts/CustomEditors/Utility.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 81c52afc12640604ea284c65ebccecdf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
