+++ Assets/Scripts/GameLoop/DamageController.cs
@@ -1,16 +1,19 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.Events;
 using UnityEngine.SceneManagement;
+
 
 public class DamageController : MonoBehaviour {
 
 	public float maxHealth;
 	private float currentHealth;
-
+	
 	private void Awake()
 	{
 		currentHealth = maxHealth;
+
 	}
 
 	public virtual void TakeDamage(float _damage)
@@ -21,5 +24,10 @@ public class DamageController : MonoBehaviour {
 			gameObject.SendMessage("Death");
 		}
 		
+	}
+
+	protected virtual void Death()
+	{
+
 	}
 }
