+++ Assets/Scripts/Utility/_Prototype.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10bfff7dd12d99e48becdbdd07851b9d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
